#6-refactor: hook 코드 분리
